requirements.txt
numpy==2.3.5
pillow==12.0.0
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
shiboken6==6.10.0
